Added kerning offset for characters
Reduced the size of the "-" character

No bitwise ANDing, but there are 4 empty columns on either side of the characters so they can be squashed together.

Can display "-0.00A" or "-000mA"
Which is enough resolution for me.

You could squeeze another character on the display if you shifted the "." over.